Allow for re-acquiring a lock from another go-routine / process.
If the value is set using WithGenValueFunc during a mutex Locking, we
should be able to use the same value when Unlocking or Extending.

The current implementation only allows a Mutex to be unlocked if the
specific instance being unlocked was the one that was originally locked.
This is because the Mutex.value is not expored, and is only ever set
when the Mutex is locked.  This change (among some house-keeping)
introduces a call to `setValue()` on each call to `LockContext`,
`UnlockContext`, and `ExtendContext`.

> Make tests not take forever
> Add a locked bool member to Mutex. This allows re-retrieval, and Extend to be called on an existing lock to "re-acquire" and be able to unlock